Update CLI
This commit remnames the "cli" package to "cmd" to better fit go's
standard package layout. In addition, flags.go file is removed, since
there is little benefit from having flag definitions in seperate files,
as there is little duplication between client and server flags. Client
flags are now defined in client.go, server flags in server.go In
addition, all files in the cmd package now correspond to particular
emmy CLI commands.
